It compiles! :)

+++ TileMap.cpp
@@ -3,7 +3,7 @@
 ////////////////////////////////////////////////////////////
 bool TileMap::load(const std::string& t_tilsetFileName,
 	const sf::Vector2u& t_tileSize,
-	const std::vector<unsigned> &t_tiles,
+	const std::vector<unsigned>& t_tiles,
 	unsigned t_width,
 	unsigned t_height)
 {
@@ -26,16 +26,16 @@ bool TileMap::load(const std::string& t_tilsetFileName,
 			auto quad{ &m_vertices[(x + y * t_width) * 4] };
 
 			// Define 4 corners
-			quad[0].position = sf::Vector2f(x * t_tileSize.x, y * t_tileSize.y);
-			quad[1].position = sf::Vector2f((x + 1) * t_tileSize.x, y * t_tileSize.y);
-			quad[2].position = sf::Vector2f((x + 1) * t_tileSize.x, (y + 1) * t_tileSize.y);
-			quad[3].position = sf::Vector2f(x * t_tileSize.x, (y + 1) * t_tileSize.y);
+			quad[0].position = sf::Vector2f(static_cast<float>(x * t_tileSize.x), static_cast<float>(y * t_tileSize.y));
+			quad[1].position = sf::Vector2f(static_cast<float>((x + 1) * t_tileSize.x), static_cast<float>(y * t_tileSize.y));
+			quad[2].position = sf::Vector2f(static_cast<float>((x + 1) * t_tileSize.x), static_cast<float>((y + 1) * t_tileSize.y));
+			quad[3].position = sf::Vector2f(static_cast<float>(x * t_tileSize.x), static_cast<float>((y + 1) * t_tileSize.y));
 
 			// Define 4 texture corners
-			quad[0].texCoords = sf::Vector2f(tex_x * t_tileSize.x, tex_y * t_tileSize.y);
-			quad[1].texCoords = sf::Vector2f((tex_x + 1) * t_tileSize.x, tex_y * t_tileSize.y);
-			quad[2].texCoords = sf::Vector2f((tex_x + 1) * t_tileSize.x, (tex_y + 1) * t_tileSize.y);
-			quad[3].texCoords = sf::Vector2f(tex_x * t_tileSize.x, (tex_y + 1) * t_tileSize.y);
+			quad[0].texCoords = sf::Vector2f(static_cast<float>(tex_x * t_tileSize.x), static_cast<float>(tex_y * t_tileSize.y));
+			quad[1].texCoords = sf::Vector2f(static_cast<float>((tex_x + 1) * t_tileSize.x), static_cast<float>(tex_y * t_tileSize.y));
+			quad[2].texCoords = sf::Vector2f(static_cast<float>((tex_x + 1) * t_tileSize.x), static_cast<float>((tex_y + 1) * t_tileSize.y));
+			quad[3].texCoords = sf::Vector2f(static_cast<float>(tex_x * t_tileSize.x), static_cast<float>((tex_y + 1) * t_tileSize.y));
 
 		}
 	return true;
